+++ tests/MetricFileExists.cpp
@@ -32,11 +32,11 @@
 //
 ///////////////////////////////////////////////////////////////////////////////
 
+#include <regex>
 #include <vector>
 
 #include <boost/algorithm/string.hpp>
 #include <boost/core/ignore_unused.hpp>
-#include <boost/regex.hpp>
 
 #include <MetricFileExists.h>
 
@@ -91,14 +91,15 @@ bool MetricFileExists::v_Test(std::istream &pStdout, std::istream &pStderr)
     for (auto it = m_fileCounts.begin(); it != m_fileCounts.end(); ++it)
     {
         int cnt = 0;
-        boost::regex r(it->first.c_str());
+        std::regex r(it->first.c_str());
 
         // Examine each file in the current path and check if it matches the
         // pattern provided. Count the number of files which match.
         for (auto &e : fs::directory_iterator(pwd))
         {
-            boost::cmatch matches;
-            if (boost::regex_match(e.path().string().c_str(), matches, r))
+            std::smatch matches;
+            std::string filename = e.path().string();
+            if (std::regex_match(filename, matches, r))
             {
                 if (matches.size() == 1)
                 {
@@ -130,11 +131,12 @@ void MetricFileExists::v_Generate(std::istream &pStdout, std::istream &pStderr)
     for (auto it = m_fileCounts.begin(); it != m_fileCounts.end(); ++it)
     {
         int cnt = 0;
-        boost::regex r(it->first.c_str());
+        std::regex r(it->first.c_str());
         for (auto &e : fs::directory_iterator(pwd))
         {
-            boost::cmatch matches;
-            if (boost::regex_match(e.path().string().c_str(), matches, r))
+            std::smatch matches;
+            std::string filename = e.path().string();
+            if (std::regex_match(filename, matches, r))
             {
                 if (matches.size() == 1)
                 {