+++ tests/MetricNoWarning.h
@@ -36,7 +36,8 @@
 #define NEKTAR_TESTS_METRICNOWARNING_H
 
 #include <Metric.h>
-#include <boost/regex.hpp>
+
+#include <regex>
 #include <vector>
 
 namespace Nektar
@@ -55,7 +56,7 @@ public:
 
 protected:
     // Regex expression that should match warning message
-    boost::regex m_regexWarning{".*WARNING.*"};
+    std::regex m_regexWarning{".*WARNING.*"};
 
     // Vector of (optional) groups
     std::vector<std::vector<std::string>> m_matches;