+++ tests/MetricRegex.h
@@ -36,7 +36,8 @@
 #define NEKTAR_TESTS_METRICREGEX_H
 
 #include <Metric.h>
-#include <boost/regex.hpp>
+
+#include <regex>
 #include <vector>
 
 namespace Nektar
@@ -78,7 +79,7 @@ public:
 
 protected:
     /// Storage for the boost regex.
-    boost::regex m_regex;
+    std::regex m_regex;
     /// Stores the multiple matches defined in each <MATCH> tag.
     std::vector<std::vector<MetricRegexFieldValue>> m_matches;
     /// If true, regex matches may be in any order in output